requirements.txt
pandas
sklearn
sqlalchemy
snowflake-sqlalchemy
matplotlib
numpy
flake8
flake8-builtins
flake8-docstrings
flake8-import-order
flake8-quotes
py2neo
